--- src/main/java/com/nhnacademy/marketgg/server/cloud/Access.java
@@ -1,19 +1,0 @@
-package com.nhnacademy.marketgg.server.cloud;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-
-@NoArgsConstructor
-@Getter
-@JsonIgnoreProperties(ignoreUnknown = true)
-@ToString
-public class Access {
-
-    private Token token;
-
-    private User user;
-
-}
